database working, controllers for creating user and getting users
