+++ .gitignore
@@ -40,3 +40,6 @@ testem.log
 # System files
 .DS_Store
 Thumbs.db
+
+# Backend Secrets
+backend/models/secrets.go
