--- api/go.mod
@@ -1,5 +1,0 @@
-module src/api
-
-go 1.19
-
-require github.com/gorilla/mux v1.8.0
