+++ backend/api.rest
@@ -1,0 +1,28 @@
+@host = localhost:8080
+
+POST http://{{host}}/signup HTTP/1.1
+content-type: application/json
+    
+{
+    "firstname": "test-firstname",
+    "lastname": "test-lastname",
+    "email": "test-email",
+    "password": "test-password"
+}
+
+###
+
+POST http://{{host}}/signup HTTP/1.1
+content-type: application/json
+    
+{
+    "firstname": "test-firstname2",
+    "lastname": "test-lastname2",
+    "email": "test-email2",
+    "password": "test-password2"
+}
+
+###
+
+GET http://{{host}}/users
+content-type: application/json
