+++ backend/go.mod
@@ -1,0 +1,15 @@
+module budget-tracker
+
+go 1.19
+
+require (
+	github.com/gorilla/mux v1.8.0
+	gorm.io/driver/mysql v1.4.5
+	gorm.io/gorm v1.24.5
+)
+
+require (
+	github.com/go-sql-driver/mysql v1.7.0 // indirect
+	github.com/jinzhu/inflection v1.0.0 // indirect
+	github.com/jinzhu/now v1.1.5 // indirect
+)
